Changed the floating point type from double to float

+++ engine-2D/include/body2D.hpp
@@ -12,24 +12,24 @@ namespace physics
         body2D() = default;
         body2D(const vec2 &pos);
         body2D(const vec2 &pos, const vec2 &vel);
-        body2D(const vec2 &pos, const vec2 &vel, double mass);
-        body2D(const vec2 &pos, const vec2 &vel, double mass, double charge);
+        body2D(const vec2 &pos, const vec2 &vel, float mass);
+        body2D(const vec2 &pos, const vec2 &vel, float mass, float charge);
 
         vec2 &pos();
         vec2 &vel();
         const vec2 &pos() const;
         const vec2 &vel() const;
-        double mass() const;
-        double charge() const;
+        float mass() const;
+        float charge() const;
 
         void pos(const vec2 &pos);
         void vel(const vec2 &vel);
-        void mass(double mass);
-        void charge(double charge);
+        void mass(float mass);
+        void charge(float charge);
 
     private:
         vec2 m_pos, m_vel;
-        double m_mass = 1.0, m_charge = 1.0;
+        float m_mass = 1.f, m_charge = 1.f;
     };
 }
