+++ vec-ptr/include/vec_ptr.hpp
@@ -10,15 +10,15 @@ namespace utils
     {
     public:
         vec_ptr() = default;
-        vec_ptr(std::vector<double> *buffer, std::size_t index = 0);
+        vec_ptr(std::vector<float> *buffer, std::size_t index = 0);
 
-        double &operator[](const std::size_t idx);
-        const double &operator[](const std::size_t idx) const;
+        float &operator[](const std::size_t idx);
+        const float &operator[](const std::size_t idx) const;
 
         explicit operator bool() const;
 
     private:
-        std::vector<double> *m_buffer = nullptr;
+        std::vector<float> *m_buffer = nullptr;
         std::size_t m_index;
     };
