+++ vector/include/vec2.hpp
@@ -10,28 +10,28 @@ namespace vec
     class vec2
     {
     public:
-        double x = 0.0, y = 0.0;
+        float x = 0.f, y = 0.f;
 
         vec2() = default;
-        vec2(double x, double y);
+        vec2(float x, float y);
         vec2(const vec3 &v3);
 
-        double dot(const vec2 &v) const;
+        float dot(const vec2 &v) const;
 
-        double sq_dist(const vec2 &v) const;
-        double dist(const vec2 &v) const;
+        float sq_dist(const vec2 &v) const;
+        float dist(const vec2 &v) const;
 
-        double sq_norm() const;
-        double norm() const;
+        float sq_norm() const;
+        float norm() const;
 
         void normalize();
         vec2 normalized() const;
 
-        void rotate(double angle);
-        vec2 rotated(double angle) const;
+        void rotate(float angle);
+        vec2 rotated(float angle) const;
 
-        double angle() const;
-        double angle(const vec2 &v) const;
+        float angle() const;
+        float angle(const vec2 &v) const;
     };
 
     vec2 operator+(const vec2 &v);
@@ -42,41 +42,41 @@ namespace vec
 
     vec2 operator-(const vec2 &lhs, const vec2 &rhs);
 
-    vec2 operator+(const vec2 &lhs, double rhs);
+    vec2 operator+(const vec2 &lhs, float rhs);
 
-    vec2 operator-(const vec2 &lhs, double rhs);
+    vec2 operator-(const vec2 &lhs, float rhs);
 
-    vec2 operator+(double lhs, const vec2 &rhs);
+    vec2 operator+(float lhs, const vec2 &rhs);
 
-    vec2 operator-(double lhs, const vec2 &rhs);
+    vec2 operator-(float lhs, const vec2 &rhs);
 
     vec2 operator+=(vec2 &lhs, const vec2 &rhs);
 
     vec2 operator-=(vec2 &lhs, const vec2 &rhs);
 
-    vec2 operator+=(vec2 &lhs, double rhs);
+    vec2 operator+=(vec2 &lhs, float rhs);
 
-    vec2 operator-=(vec2 &lhs, double rhs);
+    vec2 operator-=(vec2 &lhs, float rhs);
 
     vec2 operator*(const vec2 &lhs, const vec2 &rhs);
 
     vec2 operator/(const vec2 &lhs, const vec2 &rhs);
 
-    vec2 operator*(const vec2 &lhs, double rhs);
+    vec2 operator*(const vec2 &lhs, float rhs);
 
-    vec2 operator/(const vec2 &lhs, double rhs);
+    vec2 operator/(const vec2 &lhs, float rhs);
 
-    vec2 operator*(double lhs, const vec2 &rhs);
+    vec2 operator*(float lhs, const vec2 &rhs);
 
-    vec2 operator/(double lhs, const vec2 &rhs);
+    vec2 operator/(float lhs, const vec2 &rhs);
 
     vec2 operator*=(vec2 &lhs, const vec2 &rhs);
 
     vec2 operator/=(vec2 &lhs, const vec2 &rhs);
 
-    vec2 operator*=(vec2 &lhs, double rhs);
+    vec2 operator*=(vec2 &lhs, float rhs);
 
-    vec2 operator/=(vec2 &lhs, double rhs);
+    vec2 operator/=(vec2 &lhs, float rhs);
 
     bool operator==(const vec2 &lhs, const vec2 &rhs);
